pyproject.toml
[tool.poetry]
name = "panel"
version = "0.1.0"
description = "Linux Panel From QiYuTech"
authors = ["dev <dev@qiyutech.tech>"]

[tool.poetry.dependencies]
python = "^3.9"

# Web Framework
django = ">=3.2,<4.0"

# psutil for real time monitor
psutil = ">=5.8,<6.0"

# FTP server
pyftpdlib = ">=1.5.6,<1.6"

# for production deploy
gunicorn = ">=20.0,<21"

# for log
django-structlog = ">=2.1,<2.2"

# for oauth2 support
django-oauth-toolkit = ">=1.5,<1.6"

# OTP support
django-otp = ">1.0,<1.1"

# OTP QR image
qrcode = ">=6.1,<7"
pillow = ">=8.0,<9"

# API 类型定义
pydantic = ">=1.7,<2.0"

# API for Django
django-ninja = ">=0.13,<0.14"

# 依赖
django-qiyu-utils = ">=0.3,<0.4"
dj-qiyu-tpl = ">=0.4,<0.5"

# Docker API
docker = ">=5,<6.0"

# LDAP Server 支持
ldaptor = ">=21,<22"

[tool.poetry.dev-dependencies]
black = "*"
pylama = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
